backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# HTTP client
httpx==0.28.1
aiohttp==3.11.11

# Database
supabase==2.11.0

# OpenAI
openai==1.58.1

# XML parsing (for PubMed responses)
lxml==5.3.0

# Authentication
PyJWT==2.9.0
email-validator==2.1.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# Utilities
tenacity==9.0.0
